move ModelCriteria::useQuery() tests into own class
